Move client ui classes to client subpackage
