+++ src/main/java/ay2122s1_cs2103t_w16_2/btbb/ui/tabcontent/HomeTabContent.java
@@ -2,8 +2,8 @@ package ay2122s1_cs2103t_w16_2.btbb.ui.tabcontent;
 
 import ay2122s1_cs2103t_w16_2.btbb.model.client.Client;
 import ay2122s1_cs2103t_w16_2.btbb.model.order.Order;
-import ay2122s1_cs2103t_w16_2.btbb.ui.ClientListPanel;
 import ay2122s1_cs2103t_w16_2.btbb.ui.UiPart;
+import ay2122s1_cs2103t_w16_2.btbb.ui.client.ClientListPanel;
 import ay2122s1_cs2103t_w16_2.btbb.ui.order.OrderListPanel;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
